requirements-dev.txt
pytest>=8.0.0
psycopg[binary]>=3.1.0
requests>=2.32.0
psycopg2-binary>=2.9.9
